+++ tests/path_gateway_dag_test.go
@@ -367,8 +367,7 @@ func TestPathing(t *testing.T) {
 				Path("/ipfs/{{cid}}/foo", dagJSONTraversalCID).
 				Query("format", "dag-json"),
 			Response: Expect().
-				Status(501).
-				Body(Contains("reading IPLD Kinds other than Links (CBOR Tag 42) is not implemented")),
+				Status(501), // reading IPLD Kinds other than Links (CBOR Tag 42) is not implemented
 		},
 		{
 			Name: "GET DAG-JSON traverses multiple links",
@@ -396,8 +395,7 @@ func TestPathing(t *testing.T) {
 				Path("/ipfs/{{cid}}/foo", dagCBORTraversalCID).
 				Query("format", "dag-cbor"),
 			Response: Expect().
-				Status(501).
-				Body(Contains("reading IPLD Kinds other than Links (CBOR Tag 42) is not implemented")),
+				Status(501), // reading IPLD Kinds other than Links (CBOR Tag 42) is not implemented
 		},
 		{
 			Name: "GET DAG-CBOR traverses multiple links",
@@ -694,7 +692,6 @@ func TestNativeDag(t *testing.T) {
 		RunWithSpecs(t, rangeTests, specs.PathGatewayDAG)
 	}
 
-	var dagCborHTMLRendering []byte
 	RunWithSpecs(t, SugarTests{
 		SugarTest{
 			Name: "Convert application/vnd.ipld.dag-cbor to text/html",
@@ -704,34 +701,9 @@ func TestNativeDag(t *testing.T) {
 				Headers(
 					Header("Accept", "text/html"),
 				),
-			Response: Expect().Body(Checks("", func(t []byte) bool {
-				innerCheck := Contains("</html>").Check(string(t))
-				if innerCheck.Success {
-					dagCborHTMLRendering = t
-					return true
-				}
-				return false
-			})),
+			Response: Expect().Body(Contains("</html>")),
 		},
 	}, specs.PathGatewayDAG)
-
-	if dagCborHTMLRendering != nil {
-		rangeTests := helpers.OnlyRandomRangeTests(t,
-			SugarTest{
-				Name: "Convert application/vnd.ipld.dag-cbor to text/html with range request includes correct bytes",
-				Hint: "",
-				Request: Request().
-					Path("/ipfs/{{cid}}/", dagCborCID).
-					Headers(
-						Header("Accept", "text/html"),
-					),
-				Response: Expect(),
-			},
-			dagCborHTMLRendering,
-			"text/html")
-
-		RunWithSpecs(t, rangeTests, specs.PathGatewayDAG)
-	}
 }
 
 func TestGatewayJSONCborAndIPNS(t *testing.T) {